Implement multi-cycle instructions
Some instructions access memory at different points in their
routine, for example, a read followed by a write. These
operations are not performed atomically, but they are predictable.

This change uses the last remaining bit in zil to signal that the
current instruction is already in-progress and should be continued
instead of started at the beginning as usual.