Assume that MMU subroutines are word aligned
Because of the way that the MMU dispatches to subroutines for read and
write operations specific to the high byte of the target address, the
subroutine's location must somehow fit in 1 byte.

To make this work, I stored the relative offset to each routine from a
known base address, and use that to jump to the right spot. This works
fine, but limits the distance of the jump to 256 bytes of subroutine
code, which has run out as new features have been added to the MMU.

To alleviate this, I'm taking advantage of the fact that every
instruction is 4 bytes, making the base address of each routine aligned
to 4 bytes, which makes the offset's lower 2 bits zero. This frees up 2
higher bits effectively quadrupling the space for subroutines at no
additional cost.

If at some point more space is needed, intentionally 8-byte aligning
each subroutine's start address and applying the same technique would
again double the span, but at this point I don't think it will be
necessary.