Completely rewrite the GPU and framebuffer
The GPU now acts more like a dispatcher for the various modes. The Linux
framebuffer code was pulled out into a separate fbdev.S file, which will
represent a new 'screen' concept, separating the Linux-specific code
from the 'display' concept, which is implemented in framebuffer.S.

After this rewrite, sprite palettes are working. The display
implementation also respects the 10-sprites-per-line hardware
limitation.

As part of moving the Linux code into fbdev.S, the routines were
modified to assume a 15-bit framebuffer instead of 32-bit. This aligns
better with what the GBC produces, so hopefully it will make for less
refactoring later.

Remaining work in this space includes:
* Utilizing WX when rendering the window
* Intelligently stopping BG rendering where the window will occlude
* Proper prioritization of sprites based on X coordinate
* Window line skipping