Combine B and C registers
This reduces the CPU's allocation of the general purpose registers. Some
of the CPU instructions got a little bit less efficient, and some got
better.

Git says that there are 29 more non-empty lines in the CPU, which is
probably a good approximation of the performance cost of this change.
I'm betting that the price is worth it to free a register.