Add some basic OAM corruption and read blocks
First, make OAM reads return FF during graphics modes where the CPU does
not have access. This is necessary to run the OAM bug tests properly.

Second, implement some basic OAM bug routines. This isn't a complete
implementation in the slightest - it doesn't corrupt the correct
locations in OAM based on what the graphics block is accessing, and it
also doesn't corrupt using the right pattern based on the opcode that
triggered the event.