Improve timer on DIV write and overflow
The reset routine on DIV write was taking one less cycle than it should
have.

Additionally, there is a nuance surrounding the overflow condition where
the copy from TMA to TIMA on overflow is actually delayed one cycle,
meaning that the register reads 0x00 for a single cycle.

This isn't a very efficient approach to solving this issue, and it
should be optimized at some point.