Only ever process buttons before CPU events
We never make decisions in graphics or timer routines based on buttons,
so there is no reason to check for input when we're only processing
those routines. Since the CPU is the only other block being dispatched,
skip the button check unless the CPU is being triggered next.

In the Linux block set, the evdev routine is extremely expensive for
some reason. It makes things a lot faster to forego the unnecessary
checks.